fix: run daily scans directly in scheduled function to avoid serverless timeout
The scheduled function was calling /api/cron over HTTP, which times out
at Netlify's ~26s function limit. Extracted scan logic into lib/run-scans.ts
so the scheduled function (which runs as a 15-min background function) calls
it directly — no HTTP round-trip, no timeout.
