+++ netlify/functions/scheduled-scan.ts
@@ -1,18 +1,9 @@
 import type { Config } from "@netlify/functions";
 
+import { runScans } from "../../lib/run-scans";
+
 export default async () => {
-  const url = process.env.URL;
-  const secret = process.env.CRON_SECRET;
-  if (!url || !secret) {
-    throw new Error("URL and CRON_SECRET env vars are required");
-  }
-  const res = await fetch(`${url}/api/cron`, {
-    method: "POST",
-    headers: { "x-cron-secret": secret }
-  });
-  if (!res.ok) {
-    throw new Error(`/api/cron responded ${res.status}`);
-  }
+  await runScans();
 };
 
 export const config: Config = {
